+++ CHANGES.rst
@@ -9,3 +9,12 @@ Changelog of lizard-auth-server
 
 - First release of lizard-auth-server based on a heavily modified
   django-simple-sso.
+
+- Roles, Organisations and related data are now part of
+  lizard_auth_server.
+
+- Information about the user's roles in organisation is passed from
+  VerifyView, along with information about the user. This is ignored
+  by old versions of lizard_auth_client but can be used by a new
+  version to construct the same information at the Portal side.
+